requirements.txt
scrapy
undetected-chromedriver
selenium
pymongo
python-dotenv
